requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0 
plantuml
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy
psycopg2
asyncpg